Rearranged the code
